DFAST-87: Move BedLevelCalculator to its own file
